refactor(query): extract StreamView to query/_stream, load_output_types to output_types
Per review — these read-model helpers didn't belong in runners/_helpers:
- StreamView (self-contained over engine+query+limit) → new secator/query/_stream.py;
  __contains__ is now an O(1) keyed count-query on the compared fields instead of an
  O(N) stream+scan.
- load_output_types → secator/output_types/_base.py, and it now yields (generator) so
  callers stream instead of materializing the rehydrated list.
Update the StreamView importers (runners/_base, report) and the load_output_types test.
