+++ secator/runners/_helpers.py
@@ -213,78 +213,6 @@ def parse_extractor(extractor):
 	return _type, _field, _condition, _group_by
 
 
-class StreamView:
-	"""Lazy, streaming view over a run-scoped store query — the read-model for a run's findings.
-
-	Iterating streams the backend cursor in batches (never materializes all N); ``len()`` is an
-	indexed count; ``bool()`` is a cheap count. Dicts are rehydrated to OutputType on the fly.
-
-	ponytail: ``__contains__`` is an O(N) stream + ``==`` scan — fine for the small membership
-	checks in the integration tests; the RAM-critical paths use ``__iter__``/``__len__`` which
-	stay flat. Upgrade __contains__ to a keyed exists-query if a hot path ever needs it.
-	"""
-	def __init__(self, engine, query, batch_size=1000, limit=0):
-		self._engine = engine
-		self._query = query
-		self._batch_size = batch_size
-		self._limit = limit
-
-	def __iter__(self):
-		n = 0
-		for batch in self._engine.iterate(self._query, self._batch_size):
-			for item in load_output_types(batch):
-				yield item
-				n += 1
-				if self._limit and n >= self._limit:
-					return
-
-	def __len__(self):
-		n = self._engine.count(self._query)
-		return min(n, self._limit) if self._limit else n
-
-	def __bool__(self):
-		# Exists-check by streaming — return on the first match instead of counting the whole set
-		# (a bare count materializes on the json backend). Used by exporters' `if not items:` guard.
-		for _ in self:
-			return True
-		return False
-
-	def __contains__(self, item):
-		return any(x == item for x in self)
-
-
-def load_output_types(docs):
-	"""Rehydrate store query results (dicts) into OutputType objects.
-
-	Items already OutputType instances pass through; dicts are mapped by their ``_type``
-	to the matching class and loaded. Docs with an unknown/missing type are skipped.
-	The uuid is taken from ``_uuid`` (json/sqlite) or ``_id`` (mongodb).
-
-	Args:
-		docs (list): Store query results (dicts and/or OutputType objects).
-
-	Returns:
-		list[OutputType]: Rehydrated output types.
-	"""
-	from secator.output_types import OUTPUT_TYPES
-	by_name = {o.get_name(): o for o in OUTPUT_TYPES}
-	out = []
-	for doc in docs:
-		if isinstance(doc, dict):
-			klass = by_name.get(doc.get('_type'))
-			if not klass:
-				continue
-			item = klass.load(doc)
-			if not item._uuid:
-				item._uuid = str(doc.get('_uuid') or doc.get('_id') or '')
-			out.append(item)
-		elif hasattr(doc, '_type') and hasattr(doc, 'toDict'):
-			# Already an OutputType. Structural check (not isinstance) so a mid-suite
-			# module reload — which changes the OutputType class identity — can't drop it.
-			out.append(doc)
-	return out
-
-
 def run_scope_query(ctx):
 	"""Bound a query to the current run via the top-most present ancestry id (scan > workflow >
 	task). Each store driver mints its {type}_id at on_init (native format) and stamps it into
